MSL: Cast when a variable's type is adjusted to what was declared in the attribute
06d483459b6fd24ed7b0a439c0bbf2d9c2eba57b introduces a vertex input type change
when there's a mismatch with host declaration. This can cause further type
mismatches in the shader when the changed type cannot be automatically cast.
For example, if the shader uses a constant of the original type, the following
error can occur:

program_source:50:39: error: implicit conversions between vector types ('uint4'
(vector of 4 'unsigned int' values) and 'int4' (vector of 4 'int' values)) are
not permitted
    float4 r0 = as_type<float4>(in.v7 * int4(3));